+++ src/views/numberAnimation/index.ts
@@ -1,0 +1,3 @@
+export default {
+  title: '数值动画'
+}