+++ src/views/timepicker/index.ts
@@ -1,0 +1,3 @@
+export default {
+  title: '时间选择器'
+}